Revert "Disable purge interval"
This reverts commit 7f4362f1b04133e3116517a3d9e2e297622b03ed.

+++ src/LocalStorageManager.ts
@@ -70,11 +70,19 @@ export async function copyToLocalStorage(location: Locations, fileName: string):
  * Starts a background task that purges old files from local storage.
  */
 export function startBackgroundPurge(): void {
-  log.verbose(
-    "Local storage",
-    `Enabling background purge every ${Settings.purgeInterval} minutes for files older than ${Settings.purgeAge} minutes.`,
-  );
-  purgeOldFiles();
+  if (Settings.purgeInterval > 0) {
+    log.verbose(
+      "Local storage",
+      `Enabling background purge every ${Settings.purgeInterval} minutes for files older than ${Settings.purgeAge} minutes.`,
+    );
+    purgeOldFiles();
+  }
+  else {
+    log.verbose(
+      "Local storage",
+      `Background purge is disabled via settings.`,
+    );  
+  }
 }
 
 /**
@@ -110,7 +118,10 @@ async function purgeOldFiles(): Promise<void> {
   );
 
   log.verbose("Local storage", "Purge complete");
-  _backgroundTimer = setTimeout(purgeOldFiles, Settings.purgeInterval * _millisecondsInAMinute);
+
+  if (Settings.purgeInterval > 0 ) {
+    _backgroundTimer = setTimeout(purgeOldFiles, Settings.purgeInterval * _millisecondsInAMinute);
+  }
 }
 
 /**
